% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{f2011_12}
\alias{f2011_12}
\title{\code{f2011_12}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 18512 rows and 3762 columns.}
\source{
\url{https://meps.ahrq.gov/mepsweb/data_stats/download_data_files_detail.jsp?cboPufNumber=HC-156}
}
\usage{
f2011_12
}
\description{
MEPS Panel 16. The persons on this data set represent those who
were in the MEPS population (U.S. civilian noninstitutionalized) for all or
part of the 2011-2012 period.
}
\keyword{datasets}
