% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{f2012_13}
\alias{f2012_13}
\title{\code{f2012_13}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17923 rows and 3495 columns.}
\source{
\url{https://meps.ahrq.gov/mepsweb/data_stats/download_data_files_detail.jsp?cboPufNumber=HC-164}
}
\usage{
f2012_13
}
\description{
MEPS Panel 17. The persons on this data set represent those who
were in the MEPS population (U.S. civilian noninstitutionalized) for all or
part of the 2012-2013 period.
}
\keyword{datasets}
