% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{f2013_14}
\alias{f2013_14}
\title{\code{f2013_14}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16714 rows and 3454 columns.}
\source{
\url{https://meps.ahrq.gov/mepsweb/data_stats/download_data_files_detail.jsp?cboPufNumber=HC-172}
}
\usage{
f2013_14
}
\description{
MEPS Panel 18. The persons on this data set represent those who
were in the MEPS population (U.S. civilian noninstitutionalized) for all or
part of the 2013-2014 period.
}
\keyword{datasets}
